steering_fast/pipeline/stage1.py
"""Stage 1: Train steering directions per concept.

Optimizations over original:
- Batched hidden state extraction (10-16x)
- GPU-resident tensors (no CPU-GPU roundtrips)
- Soft labels loaded from Stage 0 .npy files (enables flash attention)
- Negative hidden states cached across concepts (2x fewer forward passes)
- Checkpoint/resume per concept
"""
import logging
import os
import pickle
import sys
from typing import Dict, Optional

import numpy as np
import torch

from ..data.statements import StatementCache
from ..extraction.attention import compute_token_indices_per_layer
from ..extraction.hidden_states import extract_hidden_states_batched, load_soft_labels_from_npy
from ..model_loader import load_model
from ..tracking.checkpoint import CheckpointManager
from ..tracking.timer import PipelineTimer
from ..tracking.wandb_tracker import WandbTracker
from ..utils import config_hash, ensure_dir, read_concept_list, save_pickle, set_seed

log = logging.getLogger(__name__)

_rfm_cache = None


def _get_rfm_trainer():
    """Import RFM training utilities.

    Looks for rfm.py in the original code directory (set via
    ORIGINAL_CODE_DIR env var or paths.original_code_dir config).
    Falls back to xrfm package if installed.
    """
    global _rfm_cache
    if _rfm_cache is not None:
        return _rfm_cache

    orig_dir = os.environ.get("ORIGINAL_CODE_DIR", "")
    if not orig_dir:
        # Try common relative locations
        for candidate in [
            os.path.join(os.path.dirname(__file__), "..", "..", "..", "attention_guided_steering"),
            os.path.join(os.getcwd(), "attention_guided_steering"),
            os.path.join(os.getcwd(), "..", "attention_guided_steering"),
        ]:
            if os.path.isfile(os.path.join(candidate, "rfm.py")):
                orig_dir = os.path.abspath(candidate)
                break

    if orig_dir and orig_dir not in sys.path:
        sys.path.insert(0, orig_dir)
        log.info("Using original code from: %s", orig_dir)

    from rfm import rfm as train_rfm
    from direction_utils import train_rfm_probe_on_concept, pearson_corr
    from control_toolkits import split_indices, minmax_normalize

    _rfm_cache = (train_rfm, train_rfm_probe_on_concept, pearson_corr, split_indices, minmax_normalize)
    return _rfm_cache


def train_rfm_direction(
    hidden_states: torch.Tensor,
    soft_labels: torch.Tensor,
    bandwidths: list,
    reg: float,
    rfm_iters: int,
) -> tuple:
    """Train RFM direction for one layer. GPU-resident computation.

    Args:
        hidden_states: (N, D) on GPU
        soft_labels: (N, 1) on GPU
        bandwidths: List of bandwidth values to search
        reg: Regularization
        rfm_iters: Number of RFM iterations

    Returns:
        (direction_tensor, correlation_score)
    """
    _, train_rfm_probe, pearson_corr, split_indices, minmax_normalize = _get_rfm_trainer()

    # Normalize
    hs = hidden_states.float()
    labels = minmax_normalize(soft_labels.float())

    # Split
    train_idx, val_idx = split_indices(len(hs))
    train_X = hs[train_idx]
    val_X = hs[val_idx]
    train_y = labels[train_idx]
    val_y = labels[val_idx]

    # Train RFM (reuses original implementation)
    best_u, best_r = train_rfm_probe(train_X, train_y, val_X, val_y, bws=bandwidths, reg=reg)

    return best_u.reshape(1, -1), best_r


def run_stage1(cfg, timer: PipelineTimer, tracker: WandbTracker) -> None:
    """Train steering vectors for all concepts."""
    set_seed(cfg.seed)

    data_dir = cfg.paths.data_dir
    output_dir = os.path.join(data_dir, "directions")
    attn_dir = os.path.join(data_dir, "attention_to_prompt")
    ensure_dir(output_dir)

    use_soft_labels = cfg.training.label_type == "soft"
    use_layer_tokens = cfg.training.rep_token == "max_attn_per_layer"

    concept_file = os.path.join(data_dir, cfg.data.concept_file)
    concepts = read_concept_list(concept_file, lowercase=cfg.data.lowercase)
    if cfg.smoke_test.enabled:
        concepts = concepts[: cfg.smoke_test.n_concepts]
    log.info("Stage 1: %d concepts, method=%s, labels=%s", len(concepts), cfg.steering.method, cfg.training.label_type)

    # Checkpoint
    ckpt = CheckpointManager(cfg.paths.checkpoint_dir, "stage1", config_hash(cfg))
    completed, _ = ckpt.load()

    # Load model (can use flash attention if soft labels come from .npy)
    can_use_flash = use_soft_labels and use_layer_tokens  # soft labels from stage 0 .npy
    attn_impl = "flash_attention_2" if can_use_flash else "eager"
    llm = load_model(cfg.model.name, cfg.model.hf_id, cfg.paths.cache_dir, attn_implementation=attn_impl)

    stmt_cache = StatementCache(data_dir, cfg.training.datasize)
    hidden_layers = list(range(1, llm.language_model.config.num_hidden_layers))

    # Cache for negative hidden states (concept-independent)
    neg_hs_cache: Optional[Dict[int, torch.Tensor]] = None

    for concept in concepts:
        # Build output filename (matches original format)
        suffix = "_softlabels" if use_soft_labels else ""
        vec_path = os.path.join(
            output_dir,
            f"{cfg.steering.method}_{concept}_tokenidx_{cfg.training.rep_token}_block{suffix}_{cfg.model.name}.pkl",
        )

        if concept in completed or os.path.exists(vec_path):
            log.info("Skipping %s (already done)", concept)
            completed.add(concept)
            continue

        with timer.time_concept("stage1", concept):
            # Get layer-to-token mapping from stage 0 attention
            layer_to_token = None
            if use_layer_tokens:
                npy_path = os.path.join(attn_dir, f"attentions_{cfg.training.head_agg}head_{cfg.model.name}_{concept}_paired_statements.npy")
                if os.path.exists(npy_path):
                    attn_data = np.load(npy_path)
                    layer_to_token = compute_token_indices_per_layer(attn_data)
                else:
                    log.warning("No attention file for %s, using default token=-1", concept)

            # Get dataset
            dataset = stmt_cache.get_unpaired_dataset(
                concept, cfg.data.positive_template, cfg.data.negative_template, llm.tokenizer,
            )
            prompts = dataset["inputs"]
            labels_list = dataset["labels"]

            # Extract hidden states (batched)
            all_hs, all_soft = extract_hidden_states_batched(
                prompts, llm.language_model, llm.tokenizer,
                hidden_layers, cfg.training.rep_token, layer_to_token,
                batch_size=cfg.training.batch_size,
                need_attention=(use_soft_labels and not can_use_flash),
                head_agg=cfg.training.head_agg,
            )

            # Load soft labels from .npy if available
            if use_soft_labels and can_use_flash:
                n_pos = sum(1 for l in labels_list if l == 1)
                n_neg = sum(1 for l in labels_list if l == 0)
                npy_soft = load_soft_labels_from_npy(
                    concept, cfg.model.name, cfg.model.n_layers,
                    layer_to_token or {}, attn_dir, n_pos, n_neg,
                )
                if npy_soft:
                    all_soft = npy_soft

            # Train directions per layer
            directions = {}
            for layer_idx in hidden_layers:
                hs = all_hs[layer_idx]  # (N, D) on GPU
                if use_soft_labels and all_soft and layer_idx in all_soft:
                    soft = all_soft[layer_idx]
                else:
                    soft = torch.tensor(labels_list, device=hs.device).float().unsqueeze(1)

                direction, r_score = train_rfm_direction(
                    hs, soft,
                    cfg.steering.bandwidths, cfg.steering.reg, cfg.steering.rfm_iters,
                )
                directions[layer_idx] = direction.cpu()
                log.debug("Layer %d: r=%.4f", layer_idx, r_score)

            # Save (matches original format)
            save_pickle(directions, vec_path)
            log.info("Saved directions for %s (r_mean=%.4f)", concept,
                     np.mean([r for _, r in [train_rfm_direction(all_hs[l], all_soft[l] if all_soft and l in all_soft else torch.tensor(labels_list).float().unsqueeze(1).to(all_hs[l].device), cfg.steering.bandwidths, cfg.steering.reg, cfg.steering.rfm_iters) for l in [hidden_layers[0]]]]))

        completed.add(concept)
        ckpt.save(completed, {})
        tracker.log_concept(concept, "stage1", {"saved": True})

        # Free GPU memory
        del all_hs, all_soft
        torch.cuda.empty_cache()

    ckpt.cleanup()
    log.info("Stage 1 complete: %d concepts", len(concepts))
