pyproject.toml
[project]
name = "steering-fast"
version = "0.1.0"
description = "Optimized LLM steering pipeline with batched inference, checkpointing, and experiment tracking"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "accelerate>=1.0",
    "bitsandbytes>=0.40",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "openai>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "tqdm",
    "wandb",
    "torchmetrics>=1.0",
    "scikit-learn>=1.3",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
xrfm = ["xrfm @ git+https://github.com/dmbeaglehole/xRFM.git@773fae8"]

[project.scripts]
steering-fast = "steering_fast.pipeline.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["steering_fast"]
